pyproject.toml
[project]
name = "rag-rbac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "groq>=1.1.2",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.30.0",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.33.0",
    "ruff>=0.15.8",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
